+++ common/hlm_communication/src/errors.rs
@@ -1,0 +1,7 @@
+pub enum HLMCommunicationError {
+    CertFileReadError(rustls::pki_types::pem::Error),
+    CertFileNotAdded(rustls::Error),
+    MissingKeyPath,
+    KeyFileReadError(rustls::pki_types::pem::Error),
+    CouldNotConfigCert,
+}